// src/api/models/signup-user-request.ts
/* tslint:disable */
/* eslint-disable */
export interface SignupUserRequest {

  /**
   * Email of the registered user
   */
  email: string;

  /**
   * First name of the registered user
   */
  firstName: string;

  /**
   * Last name of the registered user
   */
  lastName: string;

  /**
   * Password of the registered user
   */
  password: string;
}